Create countwords.py
This countword.py is to fulfill a task that given a sentence string, like "cat bat mat cat bat cat", we need to write a function that can count the appearance time of each words and output the most frequently appeared n words, n is a given variable. In case of the same appearance, words should be sorted in alphabet sequence, a comes first.